% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text_extract.R
\name{cleaning_text_columns_extract}
\alias{cleaning_text_columns_extract}
\title{This is a wrapper around a thing that i trouble shot}
\usage{
cleaning_text_columns_extract(
  df,
  our_pattern = "<bodyText>.*?<p.*?>(.*?)</p>.*?</bodyText>"
)
}
\arguments{
\item{df}{this is a df or tibble}

\item{our_pattern}{this is a regex defaults to bodytext}
}
\description{
This function uses  extract this should be used if you can't find a regex that works
}
