% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json.R
\name{extract_json}
\alias{extract_json}
\title{Wrapper for Will's Extract JSON function}
\usage{
extract_json(your_list)
}
\arguments{
\item{your_list}{is a list of json files}
}
\description{
Wrapper for Will's Extract JSON function
}
