% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text_extract.R, R/clean_text_match.R
\name{cleaning_text_columns_match}
\alias{cleaning_text_columns_match}
\title{Cleaner for the text stuff}
\usage{
cleaning_text_columns_match(
  df,
  our_pattern =
    "<bodyText><p nitf:lede=\\\\\\"true\\\\\\">\\\\s*(.*?)\\\\s*</p></bodyText></nitf:body.content>"
)

cleaning_text_columns_match(
  df,
  our_pattern =
    "<bodyText><p nitf:lede=\\\\\\"true\\\\\\">\\\\s*(.*?)\\\\s*</p></bodyText></nitf:body.content>"
)
}
\arguments{
\item{df}{is a tibble or data.frame that is the result of extract json}

\item{our_pattern}{is a regular expression to capture texts}
}
\description{
This function makes use of str_match. This is probably the function that will
the heavy lifting

This function makes use of str_match. This is probably the function that will
the heavy lifting
}
